Update: reverseeagle obsolete; forgejo migration button
This commit removes the obsolete reverseeagle link
(some of it is archived, e.g.
https://archive.today/2021.05.17-075004/https://developers.reverseeagle.org,
but not very usable - an up-to-date alternative would be great if
people could maintain it);

adds a brief explanation of where to click to use the nice forgejo
migration buttons;

adds Software Freedom Conservancy's link for reasons to [give up GitHub](https://giveupgithub.org).